Refactor data communication and DOM manipulation logic
Refactored the methods for getting data from IP Stack and Countries APIs to return the data, instead of injecting the data directly into the DOM. This change decentralizes the responsibility of accessing data and manipulating the DOM, keeping the functions cleaner and more focused on their main tasks. Additional function, main(), is added to manage the flow of data.