Update world time display and refactor code
Updated the getWorldTime function to return an object with multiple time properties. This change supports the addition of the day of the year and week number to the displayed time information. Also, code refactoring was performed for a cleaner structure, including using constants instead of let, and separating the time display functionality into a separate function "addTimeToDashboard". This provides a cleaner code architecture and improves application performance.