Add sun information to the dashboard
In this commit, the main dashboard function `addTimeToDashboard(time)` has been augmented to also include information about sunrise, sunset, solar noon, and the length of the day. This information is separately obtained and calculated in the new `getSun()` function, and afterwards passed to `addTimeToDashboard(time, sun)`.

This additional information is useful to provide users more details about sun timings and day length, hence enhancing the dashboard functionalities. Some codes on weather information which are repetitive with the provided sun information were also commented out in `getCountryData(countryCode)`. Lastly, the display styling of body in index.html was improved for better appearance.